Require Jenkins 2.387.3 or newer, not 2.387.1
Use a consistent Jenkins minimum version to match
with other plugins and with the recommendations of
https://www.jenkins.io/doc/developer/plugin-development/choosing-jenkins-baseline/

https://stats.jenkins.io/pluginversions/authentication-tokens.html shows
that only 2% of the 61500 installations of the most recent release of
the plugin are running on Jenkins prior to 2.387.3.  That's not many
users to be inconvenienced by the change, especially since there have
been security advisories for Jenkins core since the release of 2.387.1.