Test with Java 21 (#154)
* Require Jenkins 2.387.3 or newer, not 2.387.1

Use a consistent Jenkins minimum version to match
with other plugins and with the recommendations of
https://www.jenkins.io/doc/developer/plugin-development/choosing-jenkins-baseline/

https://stats.jenkins.io/pluginversions/authentication-tokens.html shows
that only 2% of the 61500 installations of the most recent release of
the plugin are running on Jenkins prior to 2.387.3.  That's not many
users to be inconvenienced by the change, especially since there have
been security advisories for Jenkins core since the release of 2.387.1.

* Use parent pom 4.74, latest parent pom

* Use most recent plugin BOM version

* Test with Java 21

Java 21 released Sep 19, 2023. We'd like to announce full support for
Java 21 in early October and would like the most used plugins to be
compiling and testing with Java 21.

The acceptance test harness and plugin bill of materials tests are already
passing with Java 21. This is a further step to improve plugin readiness
for use with Java 21 and for development with Java 21.

Java 11 will be unsupported by Eclipse Temurin and some other Java
providers in October 2024. We'll need to move past Java 11 by that time.
It does not change the supported Java version or the byte code that is
being generated.